requirements.txt
requests>=2.31.0
icalendar>=5.0.12
PyYAML>=6.0.1
python-dotenv>=1.0.1
discord.py>=2.4.0
